% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagLinearFUN.R
\name{lagLinearFUN}
\alias{lagLinearFUN}
\title{Linear Model Using Last n Lags as the Independent Variable.}
\usage{
lagLinearFUN(x, obs, lags = 3)
}
\arguments{
\item{obs}{The column under investigation.}

\item{lags}{The number of lags to go back and average over.}
}
\description{
A function that uses the average of the previous n values as the independent variable in a linear regression model.
}
\examples{
lagLinearFUN()
}
