% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lurBuild.R
\name{lurBuildFUN}
\alias{lurBuildFUN}
\title{A function for developing and refining land use regression models}
\usage{
lurBuildFUN(x, resp, method = "caret")
}
\arguments{
\item{resp}{The response column to be used in the model.}

\item{method}{The LUR development method (`caret` method using package = DEFAULT, or `escape` method using AQ standard).}
}
\description{
This function develops suitable linear models for a predictor. Before using the function, checks need to be made on the univariate direction of effects as these need to be subjectively determined.
}
\examples{
data(iris)
lurBuildFUN(iris, resp = 'Species', method = 'caret')
}
