% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentFUN.R
\name{segmentFUN}
\alias{segmentFUN}
\title{A Segmented Regression Function.}
\usage{
segmentFUN(x, obs, comparison.value, date, min.length, break.points,
  plot = FALSE, plot.title = NA)
}
\arguments{
\item{obs}{This is the numeric column under investigation.}

\item{comparison.value}{This is the data to compare the obs against.}

\item{date}{Date column.}

\item{min.length}{This is the length of the comparison value sample to draw the break points from.}

\item{break.points}{Number of break points within the data (will make break.point + 1 regressions).}

\item{plot}{Option to create a plot with the break points. Defaults to `FALSE`.}

\item{plot.title}{If plot is `TRUE` this command provides an option for giving the plot a title.}
}
\description{
A function for finding a number of linear fits within a dataset.
}
\examples{
segmentFUN()
}
