% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_change.R
\name{changingCalFUN}
\alias{changingCalFUN}
\title{Calibration Interpolation Over Time.}
\usage{
changingCalFUN(x, obs, first.offset, second.offset, first.slope, second.slope)
}
\arguments{
\item{obs}{The column under investigation.}

\item{first.offset}{The intercept from the first calibration.}

\item{second.offset}{The intercept from the second calibration.}

\item{first.slope}{The slope from the first calibration.}

\item{second.slope}{The slope from the second calibration.}
}
\description{
A function for converting the linear coefficients over time to be weighted based on its proximity to one of the calibrations. The output is the corrected obs value.
}
\examples{
changingCalFUN()
}
