% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentFUN.R
\name{segmentFUN}
\alias{segmentFUN}
\title{A Segmented Regression Function.}
\usage{
segmentFUN(x, obs, comparison.value, date, min.length, break.points,
  plot = FALSE)
}
\arguments{
\item{obs}{This is the numeric column under investigation.}

\item{comparison.value}{This is the data to compare the obs against.}

\item{date}{Date column.}

\item{min.length}{This is the length of the sample to draw the mean and variance from.}

\item{break.points}{Number of break points within the data (will make break.point + 1 regressions).}

\item{plot}{Option to create a plot with the break points. Defaults to `FALSE`.}
}
\description{
A function for finding a number of linear fits within a dataset.
}
\examples{
segmentFUN()
}
